new central simple algebras abstract type
