+++ src/Hecke.jl
@@ -559,6 +559,7 @@ include("HeckeTypes.jl")
 include("Sparse.jl")
 include("NumField/NfRel/Types.jl")
 include("AlgAss/Types.jl")
+include("CentralSimpleAlg/Types.jl")
 include("AlgAssAbsOrd/Types.jl")
 include("AlgAssRelOrd/Types.jl")
 include("Grp.jl")
